[chore] 폴더명 변경
