+++ member-service/src/main/java/com/rainbowgon/memberservice/global/redis/BookmarkRedisConfig.java
@@ -1,4 +1,4 @@
-package com.rainbowgon.memberservice.global.config.redis;
+package com.rainbowgon.memberservice.global.redis;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,19 +11,20 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
-public class ThemeSortingRedisConfig {
+public class BookmarkRedisConfig {
 
-    @Value("${spring.redis.sorting.host}")
+    @Value("${spring.redis.bookmark.host}")
     private String host;
 
-    @Value("${spring.redis.sorting.port}")
+    @Value("${spring.redis.bookmark.port}")
     private int port;
 
 //    @Value("${redis.password}")
 //    private String password;
 
-    @Bean(name = "sortingRedisConnectionFactory")
-    public RedisConnectionFactory sortingRedisConnectionFactory() {
+    // Redis 저장소와 연결
+    @Bean(name = "bookmarkRedisConnectionFactory")
+    public RedisConnectionFactory bookmarkRedisConnectionFactory() {
 
         RedisStandaloneConfiguration redisStandaloneConfiguration = new RedisStandaloneConfiguration();
         redisStandaloneConfiguration.setHostName(host);
@@ -33,12 +34,13 @@ public class ThemeSortingRedisConfig {
         return new LettuceConnectionFactory(redisStandaloneConfiguration);
     }
 
-    @Bean(name = "sortingRedisStringTemplate")
-    public RedisTemplate<String, String> sortingRedisStringTemplate(
-            @Qualifier("sortingRedisConnectionFactory") RedisConnectionFactory sortingRedisConnectionFactory) {
+    // RedisTemplate bean 생성
+    @Bean(name = "bookmarkRedisStringTemplate")
+    public RedisTemplate<String, String> bookmarkRedisStringTemplate(
+            @Qualifier("bookmarkRedisConnectionFactory") RedisConnectionFactory bookmarkRedisConnectionFactory) {
 
         RedisTemplate<String, String> redisTemplate = new RedisTemplate<>();
-        redisTemplate.setConnectionFactory(sortingRedisConnectionFactory);
+        redisTemplate.setConnectionFactory(bookmarkRedisConnectionFactory);
 
         redisTemplate.setKeySerializer(new StringRedisSerializer());
         redisTemplate.setValueSerializer(new StringRedisSerializer());
